requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
flask>=2.2
jinja2
aiofiles
python-multipart
